+++ db/migrate/20191124121827_add_description_to_guilds.rb
@@ -1,0 +1,5 @@
+class AddDescriptionToGuilds < ActiveRecord::Migration[5.2]
+  def change
+    add_column :guilds, :description, :text, null: false, default: ''
+  end
+end