+++ frontend/components_react/guilds_list/guilds_list.jsx
@@ -18,6 +18,7 @@ export default class GuildsList extends React.Component {
       fractions: [],
       worlds: [],
       guilds: [],
+      filteredGuilds: [],
       fraction: '0',
       world: '0'
     }
@@ -56,16 +57,30 @@ export default class GuildsList extends React.Component {
 
   _getGuilds() {
     let params = []
-    if (this.state.fraction !== '0') params.push(`fraction_id=${this.state.fraction}`)
-    if (this.state.world !== '0') params.push(`world_id=${this.state.world}`)
     const url = `/api/v1/guilds.json?access_token=${this.props.access_token}&` + params.join('&')
     $.ajax({
       method: 'GET',
       url: url,
       success: (data) => {
-        this.setState({guilds: data.guilds})
+        this.setState({guilds: data.guilds}, () => {
+          this._filterGuilds()
+        })
       }
     })
+  }
+
+  _filterGuilds() {
+    const state = this.state
+    let filters = {}
+    if (state.fraction !== '0') filters['fraction_id'] = parseInt(state.fraction)
+    if (state.world !== '0') filters['world_id'] = parseInt(state.world)
+    const guilds = state.guilds.filter(function(guild) {
+      for (var key in filters) {
+        if (guild[key] === undefined || guild[key] != filters[key]) return false
+      }
+      return true
+    })
+    this.setState({filteredGuilds: guilds})
   }
 
   _renderFilters() {
@@ -73,6 +88,7 @@ export default class GuildsList extends React.Component {
       <div className="filters">
         {this._renderWorldFilter()}
         {this._renderFractionFilter()}
+        <a className="btn btn-primary btn-sm" href={`${this.props.locale === 'en' ? '' : '/' + this.props.locale}/guilds/new`}>{strings.addGuild}</a>
       </div>
     )
   }
@@ -114,12 +130,12 @@ export default class GuildsList extends React.Component {
   }
 
   _renderGuilds() {
-    return this.state.guilds.map((guild) => {
+    return this.state.filteredGuilds.map((guild) => {
       return (
         <tr className="guild_link" onClick={this._goToGuild.bind(this, guild.slug)} key={guild.id}>
-          <td>{guild.id}</td>
-          <td className={guild.fraction.name.en.toLowerCase()}>{guild.name}</td>
-          <td>{guild.world.name} ({guild.world.zone})</td>
+          <td className={guild.fraction_name.en.toLowerCase()}>{guild.name}</td>
+          <td>{guild.full_name}</td>
+          <td>{guild.description}</td>
         </tr>
       )
     })
@@ -127,13 +143,13 @@ export default class GuildsList extends React.Component {
 
   _onChangeWorld(event) {
     this.setState({world: event.target.value}, () => {
-      this._getGuilds()
+      this._filterGuilds()
     })
   }
 
   _onChangeFraction(event) {
     this.setState({fraction: event.target.value}, () => {
-      this._getGuilds()
+      this._filterGuilds()
     })
   }
 
@@ -144,13 +160,14 @@ export default class GuildsList extends React.Component {
   render() {
     return (
       <div className="guilds">
+        <h2>{strings.guilds}</h2>
         {this._renderFilters()}
         <table className="table table-striped table-sm">
           <thead>
             <tr>
-              <th>{strings.id}</th>
               <th>{strings.name}</th>
               <th>{strings.world}</th>
+              <th>{strings.description}</th>
             </tr>
           </thead>
           <tbody>