+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- description for guilds
+
+### Modified
+- guilds page
+
 ## [0.8.6] - 2019-11-23
 ### Added
 - model WorldFraction for combination of worlds and fractions